Combine code into single file
It's against my training to place all of these into a single file. It's more logical to keep into their own file and, thus, their own module. However, as this is a tutorial for others who may wish to use the file as an exectuable in other repos, it makes sense to place these all in one file.
